Test the new Flows Validation Error hook
1. Break down the 'test_exception_hooks' module into a directory for
   easier future extension, at `tests/functional/exception_handling/'

2. Add a new file to that directory, for testing the flows hooks, with
   exactly one test in it

This tests the "single error message path" for the hook, ensuring that
it presents the 'message' field which was given by the API.
